+++ services/scheduler/advanced/service_test.go
@@ -74,7 +74,7 @@ func TestJob(t *testing.T) {
 		run++
 	}
 
-	require.NoError(t, s.ScheduleJob(ctx, "Test job", time.Now().Add(20*time.Millisecond), runFunc, nil))
+	require.NoError(t, s.ScheduleJob(ctx, "Test", "Test job", time.Now().Add(20*time.Millisecond), runFunc, nil))
 	require.Equal(t, 0, run)
 	time.Sleep(time.Duration(50) * time.Millisecond)
 	assert.Equal(t, 1, run)
@@ -92,7 +92,7 @@ func TestJobExists(t *testing.T) {
 		run++
 	}
 
-	require.NoError(t, s.ScheduleJob(ctx, "Test job", time.Now().Add(10*time.Second), runFunc, nil))
+	require.NoError(t, s.ScheduleJob(ctx, "Test", "Test job", time.Now().Add(10*time.Second), runFunc, nil))
 
 	require.True(t, s.JobExists(ctx, "Test job"))
 	require.False(t, s.JobExists(ctx, "Unknown job"))
@@ -113,7 +113,7 @@ func TestCancelJob(t *testing.T) {
 		run++
 	}
 
-	require.NoError(t, s.ScheduleJob(ctx, "Test job", time.Now().Add(100*time.Millisecond), runFunc, nil))
+	require.NoError(t, s.ScheduleJob(ctx, "Test", "Test job", time.Now().Add(100*time.Millisecond), runFunc, nil))
 	require.Equal(t, 0, run)
 	require.Len(t, s.ListJobs(ctx), 1)
 	require.NoError(t, s.CancelJob(ctx, "Test job"))
@@ -142,9 +142,9 @@ func TestCancelJobs(t *testing.T) {
 		run++
 	}
 
-	require.NoError(t, s.ScheduleJob(ctx, "Test job 1", time.Now().Add(100*time.Millisecond), runFunc, nil))
-	require.NoError(t, s.ScheduleJob(ctx, "Test job 2", time.Now().Add(100*time.Millisecond), runFunc, nil))
-	require.NoError(t, s.ScheduleJob(ctx, "No cancel job", time.Now().Add(100*time.Millisecond), runFunc, nil))
+	require.NoError(t, s.ScheduleJob(ctx, "Test", "Test job 1", time.Now().Add(100*time.Millisecond), runFunc, nil))
+	require.NoError(t, s.ScheduleJob(ctx, "Test", "Test job 2", time.Now().Add(100*time.Millisecond), runFunc, nil))
+	require.NoError(t, s.ScheduleJob(ctx, "Test", "No cancel job", time.Now().Add(100*time.Millisecond), runFunc, nil))
 	require.Equal(t, 0, run)
 	require.Len(t, s.ListJobs(ctx), 3)
 	s.CancelJobs(ctx, "Test job")
@@ -165,7 +165,7 @@ func TestCancelJobIfExists(t *testing.T) {
 		run++
 	}
 
-	require.NoError(t, s.ScheduleJob(ctx, "Test job", time.Now().Add(100*time.Millisecond), runFunc, nil))
+	require.NoError(t, s.ScheduleJob(ctx, "Test", "Test job", time.Now().Add(100*time.Millisecond), runFunc, nil))
 	require.Equal(t, 0, run)
 	require.Len(t, s.ListJobs(ctx), 1)
 	s.CancelJobIfExists(ctx, "Test job")
@@ -187,7 +187,7 @@ func TestCancelParentContext(t *testing.T) {
 		run++
 	}
 
-	require.NoError(t, s.ScheduleJob(ctx, "Test job", time.Now().Add(100*time.Millisecond), runFunc, nil))
+	require.NoError(t, s.ScheduleJob(ctx, "Test", "Test job", time.Now().Add(100*time.Millisecond), runFunc, nil))
 	require.Len(t, s.ListJobs(ctx), 1)
 	require.Equal(t, 0, run)
 	cancel()
@@ -207,7 +207,7 @@ func TestRunJob(t *testing.T) {
 		run++
 	}
 
-	require.NoError(t, s.ScheduleJob(ctx, "Test job", time.Now().Add(time.Second), runFunc, nil))
+	require.NoError(t, s.ScheduleJob(ctx, "Test", "Test job", time.Now().Add(time.Second), runFunc, nil))
 	require.Len(t, s.ListJobs(ctx), 1)
 	require.Equal(t, 0, run)
 	require.NoError(t, s.RunJob(ctx, "Test job"))
@@ -227,7 +227,7 @@ func TestRunJobIfExists(t *testing.T) {
 		run++
 	}
 
-	require.NoError(t, s.ScheduleJob(ctx, "Test job", time.Now().Add(time.Second), runFunc, nil))
+	require.NoError(t, s.ScheduleJob(ctx, "Test", "Test job", time.Now().Add(time.Second), runFunc, nil))
 	require.Equal(t, 0, run)
 	s.RunJobIfExists(ctx, "Unknown job")
 	require.Equal(t, 0, run)
@@ -260,7 +260,7 @@ func TestPeriodicJob(t *testing.T) {
 		return time.Now().Add(100 * time.Millisecond), nil
 	}
 
-	require.NoError(t, s.SchedulePeriodicJob(ctx, "Test periodic job", runtimeFunc, nil, runFunc, nil))
+	require.NoError(t, s.SchedulePeriodicJob(ctx, "Test", "Test periodic job", runtimeFunc, nil, runFunc, nil))
 	require.Len(t, s.ListJobs(ctx), 1)
 	require.Equal(t, 0, run)
 	time.Sleep(time.Duration(110) * time.Millisecond)
@@ -291,7 +291,7 @@ func TestCancelPeriodicJob(t *testing.T) {
 		return time.Now().Add(100 * time.Millisecond), nil
 	}
 
-	require.NoError(t, s.SchedulePeriodicJob(ctx, "Test periodic job", runtimeFunc, nil, runFunc, nil))
+	require.NoError(t, s.SchedulePeriodicJob(ctx, "Test", "Test periodic job", runtimeFunc, nil, runFunc, nil))
 	require.Len(t, s.ListJobs(ctx), 1)
 	require.Equal(t, 0, run)
 	require.NoError(t, s.CancelJob(ctx, "Test periodic job"))
@@ -315,7 +315,7 @@ func TestCancelPeriodicParentContext(t *testing.T) {
 		return time.Now().Add(100 * time.Millisecond), nil
 	}
 
-	require.NoError(t, s.SchedulePeriodicJob(ctx, "Test job", runtimeFunc, nil, runFunc, nil))
+	require.NoError(t, s.SchedulePeriodicJob(ctx, "Test", "Test job", runtimeFunc, nil, runFunc, nil))
 	require.Len(t, s.ListJobs(ctx), 1)
 	require.Equal(t, 0, run)
 	cancel()
@@ -342,7 +342,7 @@ func TestLimitedPeriodicJob(t *testing.T) {
 		return time.Now().Add(10 * time.Millisecond), nil
 	}
 
-	require.NoError(t, s.SchedulePeriodicJob(ctx, "Test job", runtimeFunc, nil, runFunc, nil))
+	require.NoError(t, s.SchedulePeriodicJob(ctx, "Test", "Test job", runtimeFunc, nil, runFunc, nil))
 	require.Len(t, s.ListJobs(ctx), 1)
 	require.Equal(t, 0, run)
 	time.Sleep(time.Duration(50) * time.Millisecond)
@@ -368,7 +368,7 @@ func TestBadPeriodicJob(t *testing.T) {
 		return time.Now().Add(10 * time.Millisecond), nil
 	}
 
-	require.NoError(t, s.SchedulePeriodicJob(ctx, "Test job", runtimeFunc, nil, runFunc, nil))
+	require.NoError(t, s.SchedulePeriodicJob(ctx, "Test", "Test job", runtimeFunc, nil, runFunc, nil))
 	require.Len(t, s.ListJobs(ctx), 1)
 	require.Equal(t, 0, run)
 	time.Sleep(time.Duration(50) * time.Millisecond)
@@ -391,14 +391,14 @@ func TestDuplicateJobName(t *testing.T) {
 		return time.Now().Add(100 * time.Millisecond), nil
 	}
 
-	require.NoError(t, s.ScheduleJob(ctx, "Test duplicate job", time.Now().Add(time.Second), runFunc, nil))
+	require.NoError(t, s.ScheduleJob(ctx, "Test", "Test duplicate job", time.Now().Add(time.Second), runFunc, nil))
 	require.Len(t, s.ListJobs(ctx), 1)
-	require.EqualError(t, s.ScheduleJob(ctx, "Test duplicate job", time.Now().Add(time.Second), runFunc, nil), scheduler.ErrJobAlreadyExists.Error())
+	require.EqualError(t, s.ScheduleJob(ctx, "Test", "Test duplicate job", time.Now().Add(time.Second), runFunc, nil), scheduler.ErrJobAlreadyExists.Error())
 	require.Len(t, s.ListJobs(ctx), 1)
 
-	require.NoError(t, s.SchedulePeriodicJob(ctx, "Test duplicate periodic job", runtimeFunc, nil, runFunc, nil))
+	require.NoError(t, s.SchedulePeriodicJob(ctx, "Test", "Test duplicate periodic job", runtimeFunc, nil, runFunc, nil))
 	require.Len(t, s.ListJobs(ctx), 2)
-	require.EqualError(t, s.SchedulePeriodicJob(ctx, "Test duplicate periodic job", runtimeFunc, nil, runFunc, nil), scheduler.ErrJobAlreadyExists.Error())
+	require.EqualError(t, s.SchedulePeriodicJob(ctx, "Test", "Test duplicate periodic job", runtimeFunc, nil, runFunc, nil), scheduler.ErrJobAlreadyExists.Error())
 	require.Len(t, s.ListJobs(ctx), 2)
 
 	require.NoError(t, s.CancelJob(ctx, "Test duplicate job"))
@@ -422,16 +422,16 @@ func TestBadJobs(t *testing.T) {
 		return time.Now().Add(100 * time.Millisecond), nil
 	}
 
-	require.EqualError(t, s.ScheduleJob(ctx, "", time.Now(), runFunc, nil), scheduler.ErrNoJobName.Error())
+	require.EqualError(t, s.ScheduleJob(ctx, "Test", "", time.Now(), runFunc, nil), scheduler.ErrNoJobName.Error())
 	require.Len(t, s.ListJobs(ctx), 0)
-	require.EqualError(t, s.ScheduleJob(ctx, "Test bad job", time.Now(), nil, nil), scheduler.ErrNoJobFunc.Error())
+	require.EqualError(t, s.ScheduleJob(ctx, "Test", "Test bad job", time.Now(), nil, nil), scheduler.ErrNoJobFunc.Error())
 	require.Len(t, s.ListJobs(ctx), 0)
 
-	require.EqualError(t, s.SchedulePeriodicJob(ctx, "", runtimeFunc, nil, runFunc, nil), scheduler.ErrNoJobName.Error())
+	require.EqualError(t, s.SchedulePeriodicJob(ctx, "Test", "", runtimeFunc, nil, runFunc, nil), scheduler.ErrNoJobName.Error())
 	require.Len(t, s.ListJobs(ctx), 0)
-	require.EqualError(t, s.SchedulePeriodicJob(ctx, "Test bad period job", nil, nil, runFunc, nil), scheduler.ErrNoRuntimeFunc.Error())
+	require.EqualError(t, s.SchedulePeriodicJob(ctx, "Test", "Test bad period job", nil, nil, runFunc, nil), scheduler.ErrNoRuntimeFunc.Error())
 	require.Len(t, s.ListJobs(ctx), 0)
-	require.EqualError(t, s.SchedulePeriodicJob(ctx, "Test bad period job", runtimeFunc, nil, nil, nil), scheduler.ErrNoJobFunc.Error())
+	require.EqualError(t, s.SchedulePeriodicJob(ctx, "Test", "Test bad period job", runtimeFunc, nil, nil, nil), scheduler.ErrNoJobFunc.Error())
 	require.Len(t, s.ListJobs(ctx), 0)
 }
 
@@ -450,7 +450,7 @@ func TestManyJobs(t *testing.T) {
 
 	jobs := 2048
 	for i := 0; i < jobs; i++ {
-		require.NoError(t, s.ScheduleJob(ctx, fmt.Sprintf("Job instance %d", i), runTime, runFunc, nil))
+		require.NoError(t, s.ScheduleJob(ctx, "Test", fmt.Sprintf("Job instance %d", i), runTime, runFunc, nil))
 	}
 	require.Len(t, s.ListJobs(ctx), jobs)
 
@@ -484,13 +484,13 @@ func TestListJobs(t *testing.T) {
 	jobs := s.ListJobs(ctx)
 	require.Len(t, jobs, 0)
 
-	require.NoError(t, s.ScheduleJob(ctx, "Test job 1", time.Now().Add(time.Second), runFunc, nil))
+	require.NoError(t, s.ScheduleJob(ctx, "Test", "Test job 1", time.Now().Add(time.Second), runFunc, nil))
 
 	jobs = s.ListJobs(ctx)
 	require.Len(t, jobs, 1)
 	require.Contains(t, jobs, "Test job 1")
 
-	require.NoError(t, s.ScheduleJob(ctx, "Test job 2", time.Now().Add(time.Second), runFunc, nil))
+	require.NoError(t, s.ScheduleJob(ctx, "Test", "Test job 2", time.Now().Add(time.Second), runFunc, nil))
 
 	jobs = s.ListJobs(ctx)
 	require.Len(t, jobs, 2)
@@ -523,7 +523,7 @@ func TestLongRunningPeriodicJob(t *testing.T) {
 	}
 
 	// Schedule the job.
-	require.NoError(t, s.SchedulePeriodicJob(ctx, "Test long running periodic job", runtimeFunc, nil, jobFunc, nil))
+	require.NoError(t, s.SchedulePeriodicJob(ctx, "Test", "Test long running periodic job", runtimeFunc, nil, jobFunc, nil))
 	require.Len(t, s.ListJobs(ctx), 1)
 
 	// Sleep for 400 ms.  Expect two runs (50+100+50+100+50).
@@ -549,8 +549,8 @@ func TestOverlappingJobs(t *testing.T) {
 	}
 
 	now := time.Now()
-	require.NoError(t, s.ScheduleJob(ctx, "Test job 1", now.Add(100*time.Millisecond), jobFunc, nil))
-	require.NoError(t, s.ScheduleJob(ctx, "Test job 2", now.Add(200*time.Millisecond), jobFunc, nil))
+	require.NoError(t, s.ScheduleJob(ctx, "Test", "Test job 1", now.Add(100*time.Millisecond), jobFunc, nil))
+	require.NoError(t, s.ScheduleJob(ctx, "Test", "Test job 2", now.Add(200*time.Millisecond), jobFunc, nil))
 	require.Len(t, s.ListJobs(ctx), 2)
 
 	// Sleep to let jobs complete.
@@ -572,7 +572,7 @@ func TestMulti(t *testing.T) {
 	jobFunc := func(ctx context.Context, data interface{}) {
 		atomic.AddUint32(&run, 1)
 	}
-	require.NoError(t, s.ScheduleJob(ctx, "Test job", time.Now().Add(10*time.Second), jobFunc, nil))
+	require.NoError(t, s.ScheduleJob(ctx, "Test", "Test job", time.Now().Add(10*time.Second), jobFunc, nil))
 	require.Len(t, s.ListJobs(ctx), 1)
 
 	// Create a number of runners that will try to start the job simultaneously.
@@ -619,15 +619,20 @@ func TestCancelWhilstRunning(t *testing.T) {
 		return time.Now().Add(50 * time.Millisecond), nil
 	}
 
-	require.NoError(t, s.SchedulePeriodicJob(ctx, "Test periodic job", runtimeFunc, nil, runFunc, nil))
+	// Job takes 50 ms and runs every 50ms for a total of 100ms per tick.
+	require.NoError(t, s.SchedulePeriodicJob(ctx, "Test", "Test periodic job", runtimeFunc, nil, runFunc, nil))
 	require.Len(t, s.ListJobs(ctx), 1)
+	require.Contains(t, s.ListJobs(ctx), "Test periodic job")
 	require.Equal(t, 0, run)
 	time.Sleep(time.Duration(60) * time.Millisecond)
 	require.Equal(t, 0, run)
+	// Cancel occurs during first run.
 	require.NoError(t, s.CancelJob(ctx, "Test periodic job"))
 	require.Len(t, s.ListJobs(ctx), 0)
-	time.Sleep(time.Duration(40) * time.Millisecond)
+	// Wait for first run to finish.
+	time.Sleep(time.Duration(60) * time.Millisecond)
 	assert.Equal(t, 1, run)
+	// Ensure second run never happens.
 	time.Sleep(time.Duration(120) * time.Millisecond)
 	assert.Equal(t, 1, run)
 }